[ansible_runner] write_password_file -> write_config_files
Changes `Ansible::Runner::Credential#write_password_file` to
`Ansible::Runner::Credential#write_config_file` since this allows it to
be more generic and have this method support multiple file types.

For example, machine credential not only writes a `env/password` file,
but also a `env/ssh_key` file.  But in some of the cloud credentials,
there are cloud specific files that are generated, so this allows the
same interface to be reused, and then the logic to be implemented in
private subclass specific methods.

Machine credential has been updated to show how that implementation can
be used.