Add become_enabled option to Ansible::Runner
This is implemented as a separate option for services and automate
methods so this commit moves the command line logic from the
credential to the main Runner module.

This commit also changes the password file because the actual
prompt for the BECOME password is "BECOME password[defaults to SSH password]:"
which didn't match the original regex.