Don't restrict the zone to the embedded ansible manager zone for CRUD
There's nothing special about the zone with the appliance that has the
embedded ansible role. If we're targeting the role we should target it
without regard for zone since it's region scoped.

If we leave embedded ansible only usable on one appliance at a time
then this is an unnecessary restriction as the manager zone is set
to the server's zone when the role is assigned.

If we add the ability to use multiple servers then this will ensure
that work is only completed in the zone where the server was assigned
the embedded ansible role most recently. Which will be a very difficult
to diagnose bug.