Add Spec::Support::FakeAnsibleRepo
Adds a spec helper class which is used for quickly generating fake
ansible project repos of varying complexity that can be used to validate
functionality of `EmbeddedAnsible` classes.  Specifically will be used to
handle testing playbooks that live in nested directories, and allow for
testing against different project structures with little overhead in the
specs to define each project.

Also replaces the setup code for creating a local git repo currently in:

  embedded_ansible/automation_manager/configuration_script_source_spec.rb

Testing without MIQ
-------------------

This class is designed to be standalone, so you should be able to run
something like this:

    $ cat test_fake_ansible_repo.rb
    require_relative "./spec/support/fake_ansible_repo.rb"

    repo_path      = File.expand_path "./test_repo", __dir__
    file_structure = %w[
      roles/foo/defaults/main.yml
      roles/foo/meta/main.yml
      roles/foo/tasks/main.yml
      requirements.yml
      playbook.yml
    ]

    fake = Spec::Support::FakeAnsibleRepo.new repo_path, file_structure
    fake.generate
    fake.git_branch_create "foo"
    fake.git_branch_create "bar"

    $ ruby test_fake_ansible_repo.rb
    $ git -C ./test_repo log --stat
    commit 21848b88ce82676f928cf1145d91398423d055cd (HEAD -> master, foo, bar)
    Author: Nick LaMuro
    Date:   Wed Jul 31 20:50:43 2019 -0500

    Initial Commit

     playbook.yml                | 6 ++++++
     requirements.yml            | 5 +++++
     roles/foo/defaults/main.yml | 4 ++++
     roles/foo/meta/main.yml     | 8 ++++++++
     roles/foo/tasks/main.yml    | 6 ++++++
     5 files changed, 29 insertions(+)
    $ git -C ./test_repo branch
      foo
      bar
    * master

And confirm everything is working and created as expected.