Pass array to references()
We are now passing an array of table names into references.

We were passing a nested hash of relations into references which is not
valid.

This is undefined behavior, and not guaranteed to work in the future.
Currently, tails treats this as "just use references for all".

We are introducing includes_to_references.
It  converts an includes compatible (hash of references)
to a references compatible (array of table names)