[EmbeddedAnsible] Force embedded_ansible role for workflow
The `queue_signal` method in `AnsibleRunnerWorkflow` forces the
"ems_operations" role when it schedules a queue item, but this class
(which inherits from `AnsibleRunnerWorkflow`) is getting assigned this
job from a wrapper job that requires the `embedded_ansible` role.  In
addition, the previous job queue an job that is locked to the existing
server guid, so it is possible for that server to take the first job,
but not the second when it doesn't have both an "embedded_ansible" and
"ems_operations" role.

When a server exists that only has the "embedded_ansible" role, it is
possible to get into a state where a playbook can be scheduled, but then
is never ran because no server matches.  This fix simply always uses the
"embedded_ansible" role for everything, but tries to only modify the
lower level classes to achieve that.

Fixes:  https://bugzilla.redhat.com/show_bug.cgi?id=1742839