[EmbeddedAnsible] Fix job_plays for API
The Service UI makes use of the API to fetch both `.stdout` and
`.job_plays` together in a single call.  Without this definition,
`.job_plays` is not recognized as a valid attribute to fetch, and the
following error is returned from the API call by the ServiceUI:

    {
      "error": {
        "kind":"bad_request",
        "message":"Invalid attributes specified: job_plays",
        "klass":"Api::BadRequestError"
      }
    }

This fixes that error.

Partially fixes https://bugzilla.redhat.com/show_bug.cgi?id=1753721