CloudNetwork, NetworkPort and SecurityGroup belong to a ResourceGroup
With this commit we model a `belongs_to` relationship to a
ResourceGroup for CloudNetwork, NetworkPort and SecurityGroup models.
A ResourceGroup now has a `has_many` relationship to these network
entities. This mirrors the existing relationship between ResourceGroup
and VMs. When a resource group is deleted, references to the associated
resources are nullified.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>