remove unneeded interval check
perf_capture_queue is an internal method now.
So there parameters do not need to be checked.

The interval is determined from perf_target_to_interval_name, which
is 100% under our control, and the options hash, which is from
perf_capture_gap or perf_capture_realtime_now. (and the last one
is slated to go away soon)

Checking on every insert seemed like waste