Remove seed on rails server start in development.
In a production setting, evmserverd will be resposible for seeding the
database.  In a development setting, we should not be seeding when we
start the UI.  Developers should instead be encouraged to use bin/update
or call db:seed directly when changes are made.  Since most developers
need to call bin/update before starting a rails server, in order to
build the UI assets, this is part of the normal workflow.

This change should eliminate about 1 second at rails server boot in
development.