Raise exception if nested file will never be found
It is possible to give autoload an invalid path, and that issue will
never be found until the associated constant is vivified.  This causes
issues where people forget to remove the require_nested calls when they
remove the files.  Now, it's not possible to forget, because an
exception will be raised as soon as the parent file is loaded.