Remove MiqServer capabilities column and #update_capabilities
The only information in the capabilities column were the smart
proxy worker count which was removed in the previous commit and
whether or not the server supports VixDiskLib.

Instead of having a serialized column we can have a single boolean
for vix disk support. Additionally, because it isn't dependent on the
contents of settings we can set the value at server start rather
than every time we activate the config