Make the spec for `workers_quiesced?` make a bit more sense
Previously this spec wasn't executing the meat of any of the methods
in the loop of worker types in `workers_quiesced?` because the
worker it was creating was not in the class list.

To make this spec a bit more meaningful we need to create a worker
that actually exists in the list of worker types and seed the worker
types table.

But, after doing this the spec fails, so I changed it to always return
that the worker was stopped (it didn't really make sense otherwise)
and also removed an allow_any_instance_of in the process