Make killing container workers do the right thing
Prior to this change we would try to actually kill the worker's
pid on the orchestrator. That's never going to do what we want
and at worst could kill something we don't want to, so this change
implements MiqWorker#kill_process as #destroy_container_objects
when we're running as the orchestrator.

This isn't really different from what #stop would do, but it's better
than what we currently have.